requirements.txt
groq>=0.15.0
jsonschema==4.21.1
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
